refactor: re-arranged project structure
